requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-ollama==0.1.0
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0
sentence-transformers==3.0.1